fun_c w01 prepped, need to have a look at the textbook

+++ fun_c/wk01/provided_code/F01b.c
@@ -16,7 +16,9 @@ int main(void)
 
 	printf("Enter two integers> ");
 	scanf("%d%d", &first, &second);
-	ans = first / second;
+
+	ans = (double)first / second;
 	printf("The result is %.3lf\n", ans);
+
 	return (0);
 }
